Fixed lots of errors and redesigned disassembler

+++ Base.hpp
@@ -1,8 +1,9 @@
 #pragma once
+#include <typeinfo>
 #include <type_traits>
 #include <functional>
 #include <iostream>
-#include <typeinfo>
+#include <fstream>
 #include <memory>
 #include <vector>
 #include <set>
@@ -10,7 +11,7 @@
 
 typedef unsigned char UInt8; typedef char Int8;
 typedef unsigned short int UInt16; typedef short int Int16;
-typedef unsigned long int UInt32; typedef long int Int32; typedef float Float32;
+typedef unsigned int UInt32; typedef int Int32; typedef float Float32;
 typedef unsigned long long int UInt64; typedef long long int Int64; typedef double Float64;
 
 namespace std {
